Reusable CLI arguments
